pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "menglong"
version = "0.1.0"
description = "MengLong Agent SDK - 一个用于创建Agent的Python库"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "boto3>=1.38.0",
    "fastapi-poe>=0.0.60",
    "httpx[socks]>=0.28.1",
    "openai>=1.75.0",
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
    "rich>=13.7.0",
    "trio>=0.30.0",
]

[project.urls]
Documentation = "https://github.com/username/menglong#readme"
Issues = "https://github.com/username/menglong/issues"
Source = "https://github.com/username/menglong"

[tool.hatch.build.targets.wheel]
packages = ["src/menglong"]
